+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Detect the new format of native extensions.
 * Unblock platform classification on a new class of errors.
+* Changed the meaning of version fields in `Maintenance`:
+  * `isExperimentalVersion` now means pre-V1.
+  * `isPreReleaseVersion` now means there is a pre-release flag
+    like `-beta`, `-alpha`, etc.
 
 ## 0.7.3+1